fptr.c test: Don't use size_t without declaring it.
I want to demonstrate a hack that fixes all the other problems with this
test, and I don't want the size_t blocking the compilation.

+++ tests/checked_cases/fptr.c
@@ -9,6 +9,6 @@ void foo(void) {
   // int (*mapfunc)(unsigned int) = 0;
 }
 
-_Itype_for_any(T) void json_set_allocation_functions(_Ptr<void* (size_t s) : itype(_Array_ptr<T>) byte_count(s)> malloc,
+_Itype_for_any(T) void json_set_allocation_functions(_Ptr<void* (int s) : itype(_Array_ptr<T>) byte_count(s)> malloc,
     _Ptr<void (void* : itype(_Array_ptr<T>) byte_count(0))> free);
